[native] Fix media-cache "directory not found" warnings
Summary:
Fixes [[ https://linear.app/comm/issue/ENG-3620/warnings-about-encrypted-image-cache-directory-not-existing | ENG-3620 ]].

When clearing cache, removing an already removed directory would cause a warning to be logged. Also fixed one more place where a similiar warn could appear.

Test Plan:
The easiest way to test this is to use the iOS Simulator:
- Console-logged cache directory and opened its parent dir in Finder. Manually added some content there (it's also possible to simply open encrypted media).
- Logged-out - observed the directory content being removed.
- Logged in again. Manually deleted the dir in finder.
- Logged out again - empty dir was created.

Reviewers: ashoat, atul, kamil

Reviewed By: ashoat

Subscribers: tomek

Differential Revision: https://phab.comm.dev/D7583

+++ native/media/media-cache.js
@@ -41,7 +41,11 @@ async function getCacheSize() {
 }
 
 async function hasURI(uri: string): Promise<boolean> {
-  return await fs.exists(pathFromURI(uri));
+  const path = pathFromURI(uri);
+  if (!path) {
+    return false;
+  }
+  return await fs.exists(path);
 }
 
 async function getCachedFile(holder: string) {
@@ -57,7 +61,12 @@ async function getCachedFile(holder: string) {
 }
 
 async function clearCache() {
-  await fs.unlink(cacheDirectory);
+  const cacheDirExists = await fs.exists(cacheDirectory);
+  if (cacheDirExists) {
+    await fs.unlink(cacheDirectory);
+  }
+  // recreate empty directory
+  await ensureCacheDirectory();
 }
 
 const dataURLRegex = /^data:([^;]+);base64,([a-zA-Z0-9+/]+={0,2})$/;
